requirements.txt
absl_py==1.4.0
astropy==5.1
astunparse==1.6.3
av==11.0.0
beartype==0.17.2
boto3==1.24.28
botocore==1.27.59
clearml==1.14.4
colorama==0.4.6
conda==4.3.16
coremltools==7.1
csbdeep==0.7.4
deeplearning==0.0.1
defusedxml==0.6.0
defusedxml==0.7.1
dill==0.3.7
dill==0.3.6
dp==1.00
einops==0.7.0
etcd==2.0.8
expecttest==0.2.1
fb==0.4.0
foo==.1
fsspec==2023.4.0
fsspec==2023.10.0
fsspec==2023.9.2
fsspec==2023.3.0
google_auth_oauthlib==1.0.0
google_auth_oauthlib==1.1.0
google_auth_oauthlib==1.2.0
hypothesis==6.99.6
ignite==1.1.0
imageio==2.34.0
imageio==2.31.1
imageio==2.33.0
imageio==2.31.4
imageio==2.26.0
imread==0.7.5
itk==5.3.0
itk_core==5.3.0
itk_filtering==5.3.0
itk_io==5.3.0
itk_numerics==5.3.0
itk_registration==5.3.0
itk_segmentation==5.3.0
lazy_loader==0.3
lmdb==1.4.1
lxml==4.9.3
lxml==4.9.2
matplotlib==3.7.2
matplotlib==3.5.3
matplotlib==3.8.2
matplotlib==3.5.0
matplotlib==3.8.0
matplotlib==3.7.1
monkeytype==23.3.0
moviepy==1.0.3
networkx==3.2.1
networkx==3.0
networkx==3.1
nibabel==5.2.1
nnunetv2==2.3.1
numpydoc==1.5.0
onnxscript==0.1.0.dev20240315
osgeo==0.0.1
packaging==24.0
pooch==1.7.0
pooch==1.8.0
pooch==1.4.0
pyamg==5.0.1
pycocotools==2.0.7
pydicom==2.4.4
pydot==2.0.0
pynvml==11.5.0
pytest==7.4.0
pytorch_lightning==2.2.1
PyWavelets==1.4.1
pywin32==306
pywin32==305.1
safetensors==0.4.2
scikits.talkbox==0.2.5
sentencepiece==0.2.0
setuptools==65.6.3
setuptools==68.2.2
setuptools==68.0.0
sympy==1.12
sympy==1.11.1
tabulate==0.9.0
tabulate==0.8.10
tensorboard_data_server==0.7.2
tensorboard_plugin_profile==2.15.1
tensorboard_plugin_wit==1.8.1
tensorboardX==2.6.2.2
tensorflow_io==0.36.0
tifffile==2024.2.12
tifffile==2023.8.12
tifffile==2023.9.26
tifffile==2023.4.12
tifffile==2021.7.2
torch==2.2.1+cu118
torch==2.1.1
torch==2.1.0+cu118
torch==2.1.1+cu118
torch==1.12.1
torch==2.1.0
torch==2.0.1+cu117
torch_xla==2.2.0
torcharrow==0.1.0
torchrec==0.6.0
tqdm==4.62.3
tqdm==4.66.1
tqdm==4.65.0
traitlets==5.14.2
transformers==2.1.1
treelib==1.7.0
treelib==1.6.1
typing_extensions==4.10.0
vtk==9.2.6
xmlrunner==1.7.7
z3==0.2.0
zarr==2.17.1
